agent-hub/openai_chat_agent/pyproject.toml
#openai
[tool.poetry]
name = "openai_chat_agent"
version = "0.1.0"
authors = [
    "daiyn2002@outlook.com",
]
description = "An OpenAI Chat Agent for MoFA"
license = "MIT"
homepage = "https://github.com/your-org/openai_chat_agent"
readme = "README.md"
packages = [{ include = "openai_chat_agent" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
openai = "*"
python-dotenv = "*"

[tool.poetry.scripts]
openai_chat_agent = "openai_chat_agent.main:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
